chore: remove warning state

+++ sensing/autoware_image_diagnostics/include/autoware/image_diagnostics/image_diagnostics_node.hpp
@@ -71,10 +71,10 @@ private:
     int num_blocks_vertical;
 
     // Blockage threshold
-    int blockage_region_error_threshold;
+    double blockage_region_error_threshold;
     float blockage_ratio_threshold;
     int blockage_intensity_threshold;
-    float blockage_frequency_ratio_threshold;
+    double blockage_frequency_ratio_threshold;
 
     // Shadow clipping threshold
     int shadow_region_error_threshold;
@@ -86,10 +86,10 @@ private:
 
     // Low visibility threshold
     int low_visibility_region_error_threshold;
-    float low_visibility_frequency_threshold;
+    double low_visibility_frequency_threshold;
 
     bool use_twist;
-    float velocity_threshold;
+    double velocity_threshold;
   } params_;
 
   struct RegionFeatures
